Move, update, complete fake symbol list
