+++ cargo-rustler/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "rustler_tool"
+name = "cargo-rustler"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
 clap = { version = "4.5", features = [ "derive" ] }
 libloading = "0.8"
-rustler_sys = { version = "2.4.0", path = "../rustler_sys" }
+rustler = { version = "0.36.1", path = "../rustler" }
